move util functions to seperate files
