+++ vinoa-web/app/profile/page.tsx
@@ -1,4 +1,3 @@
-import { getUser } from '@/sanity/sanity-utils';
 import UserProfile from '@/components/UserProfile';
 
 const Profile = () => {
